+++ src/landscape/mod.rs
@@ -1,6 +1,7 @@
 #[cfg(test)]
 mod tests;
 
+mod coordinate_point;
 mod despawn_landscapes;
 mod height_map;
 mod init_height_map;
@@ -10,8 +11,8 @@ mod spawn_landscapes;
 mod spawn_rails;
 
 use bevy::prelude::*;
+pub use coordinate_point::CoordinatePoint;
 pub use height_map::HeightMap;
-use serde::{Deserialize, Serialize};
 
 const TRIANGLE_SIZE: i32 = 10;
 const LANDSCAPE_SIZE: i32 = 1000;
@@ -39,77 +40,8 @@ impl Default for Landscape {
     }
 }
 
-type PathId = (i64, i64);
-
-// TODO: merge with CoordinatePoint
-
 #[derive(Resource)]
 pub struct OriginOffset(pub CoordinatePoint);
-
-// TODO: add some basic math operations
-
-// TODO: move to file to test!
-#[derive(Default, Debug, Deserialize, Serialize, Copy, Clone, PartialEq)]
-pub struct CoordinatePoint(pub f64, pub f64);
-
-impl From<CoordinatePoint> for Vec2 {
-    fn from(val: CoordinatePoint) -> Self {
-        Vec2::new(val.0 as f32, val.1 as f32)
-    }
-}
-
-impl CoordinatePoint {
-    fn sector_coordinates(&self) -> (i64, i64) {
-        (
-            (self.0 / LANDSCAPE_SIZE as f64 + 0.5).floor() as i64,
-            (self.1 / LANDSCAPE_SIZE as f64 + 0.5).floor() as i64,
-        )
-    }
-
-    fn floor(&self) -> Self {
-        Self(self.0.floor(), self.1.floor())
-    }
-}
-
-impl std::ops::Sub<CoordinatePoint> for CoordinatePoint {
-    type Output = Self;
-
-    fn sub(self, rhs: Self) -> Self {
-        Self(self.0 - rhs.0, self.1 - rhs.1)
-    }
-}
-
-impl std::ops::Add<CoordinatePoint> for CoordinatePoint {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self {
-        Self(self.0 + rhs.0, self.1 + rhs.1)
-    }
-}
-
-impl std::ops::Add<f64> for CoordinatePoint {
-    type Output = Self;
-
-    fn add(self, rhs: f64) -> Self {
-        Self(self.0 + rhs, self.1 + rhs)
-    }
-}
-
-impl std::ops::Div<f64> for CoordinatePoint {
-    type Output = Self;
-
-    fn div(self, rhs: f64) -> Self {
-        Self(self.0 / rhs, self.1 / rhs)
-    }
-}
-
-impl std::ops::Mul<f64> for CoordinatePoint {
-    type Output = Self;
-
-    fn mul(self, rhs: f64) -> Self {
-        Self(self.0 * rhs, self.1 * rhs)
-    }
-}
 
 pub const RAIL_HEIGHT: f32 = 0.2;
 const RAIL_DISTANCE: f32 = 1.435;
